refactor(stream): move backfill/upstream_table to backfill/cdc/upstream_table
