+++ src/stream/src/executor/backfill/cdc/upstream_table/external.rs
