+++ src/stream/src/executor/backfill/mod.rs
@@ -15,5 +15,4 @@
 pub mod arrangement_backfill;
 pub mod cdc;
 pub mod no_shuffle_backfill;
-pub mod upstream_table;
 pub mod utils;
