+++ src/stream/src/from_proto/stream_cdc_scan.rs
@@ -23,8 +23,7 @@ use risingwave_pb::stream_plan::StreamCdcScanNode;
 
 use super::*;
 use crate::common::table::state_table::StateTable;
-use crate::executor::external::ExternalStorageTable;
-use crate::executor::CdcBackfillExecutor;
+use crate::executor::{CdcBackfillExecutor, ExternalStorageTable};
 
 pub struct StreamCdcScanExecutorBuilder;
